#1866, #1711, refactor
