+++ Diffing_Engine/Diffing_Engine.csproj
@@ -70,23 +70,25 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Compute\DiffWithCustomId.cs" />
+    <Compile Include="Compute\DiffRevisions.cs" />
     <Compile Include="Convert\TryParseObjectToGuid.cs" />
     <Compile Include="Create\StreamPointer.cs" />
     <Compile Include="Objects\EqualityComparers\DiffingHashComparer.cs" />
-    <Compile Include="Compute\Diffing.cs" />
-    <Compile Include="Compute\DiffingHash.cs" />
+    <Compile Include="Compute\Diff.cs" />
+    <Compile Include="Compute\Hash\DiffingHash.cs" />
     <Compile Include="Convert\ToDiffingByteArray.cs" />
     <Compile Include="Create\Delta.cs" />
     <Compile Include="Create\DiffConfig.cs" />
     <Compile Include="Create\Revision.cs" />
     <Compile Include="Modify\RemoveDuplicatesByHash.cs" />
-    <Compile Include="Modify\PrepareForDiffing.cs" />
+    <Compile Include="Modify\PrepareForRevision.cs" />
     <Compile Include="Modify\SetHashFragment.cs" />
     <Compile Include="Query\ListModifiedProperties.cs" />
     <Compile Include="Query\DifferentProperties.cs" />
     <Compile Include="Query\AnyDuplicateByHash.cs" />
     <Compile Include="Query\GetHash.cs" />
-    <Compile Include="Compute\SHA256Hash.cs" />
+    <Compile Include="Compute\Hash\SHA256Hash.cs" />
     <Compile Include="Convert\ToDiffingJson.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
